refactor MPdetector to separate module
